requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
fastdtw==0.3.4
h5py==3.11.0
hiredis==2.3.2
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
llvmlite==0.42.0
loguru==0.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
redis==5.0.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.2
tslearn==0.6.3
tzdata==2024.1
wcwidth==0.2.13
